+++ tests/TestCase/Controller/PostsControllerTest.php
@@ -216,7 +216,7 @@ class PostsControllerTest extends ControllerTestCase
         $this->assertResponseCode(302);
         $this->assertStringStartsWith('/login', $this->_response->getHeader('Location')[0]);
 
-        $this->setUserId(1);
+        $this->setAuthData();
         $this->get($url);
         $this->assertResponseOkAndNotEmpty();
         $this->assertTemplate('Posts' . DS . 'view.php');