+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.31 branch
 ### 2.31.0
+* the `ControllerTestCase` has been completely revised and some methods have been moved to `AdminControllerTestCase`,
+  (specific for controller admin), instead other methods have been removed;
 * the `AuthComponent` is deprecated. Use instead `\Authentication\Controller\Component\AuthenticationComponent`;
 * the `AuthHelper` is deprecated.  Use instead `\MeCms\View\Helper\IdentityHelper`.